Prompts now work, are in the proper place, and are hyperlinked.
This required changes to mds-output and mds-showstat;
mostly the former.  Not sure I like the look of hyperlinked
prompts---busy. An alternative is to bind a mouse key,
or modifier, in the output buffer.  Of course, we don't have
to use the hyperlink font.  Better make the chosen font
a customization.