Removed `mds-show-args-on-entry' variable and cleaned mds-showstat-update.
There was a timing problem with displaying the args that causes the
debugger to 'continue' to the end.  May revisit later, but the code
is significantly cleaner without this, and it didn't work, so the
decision was easy.

Stopped redundant calls to showstat when on line 1.

In thinking about this, it might be worthwhile to cache showstat
strings.  There is a danger in that, during development, procedures
change frequently, so the cache would have to be cleared whenever that
occurs.  A reasonable solution is to have a separate cache for each
process.  The process will almost always be restarted, so when that
happens the cache would cleared.