Use address of procname, not procedure.
Do not yet know why, but using the address to the procedure wreaks
havoc with the kernel debugger.  Maybe the pointer is changed so that
it points to a debugged version?  Will investigate later.

Am now planning to modify so that the address is prepended to the
output, rather than appearing after it.  Maybe separated by a \n.
That should be easier to handle with the regexes, and doesn't require
building new strings since it can be sent separately.  Alas, this
requires touching a fair amount of code, so will do on another branch.