Trace mode works, but does not echo statement
(mds-output-goto-source-click): changed with-current-buffer
to set-buffer.  The latter is cheaper and the result is the
same since the subsequent response from Maple causes the
more output to be displayed...

(mds-output-goto-source-line): removed save-excursion;
the reason is the same as above.

mds-regexps.el: Removed dependency on maplev.  The problem is that the
\\< (word beging), which is part of the maplev symbol regexp,
has issues.  It doesn't allow _ unless which change syntax table.
Using \\_< doesn't help because it doesn't work when symbols
are connected, say with `:-'.