Stop echoing of "stopping" when a process quits.
Rather than appearing in the output window, where it is intended to go,
the text frequenty is inserted into a random buffer.  Presumably that
happens because the buffer gets closed before the writing is complete.
This is a workaround for now.

The proper solution may be to do something in the sentinel.