Client grouping is now working.
Groups are recomputed each time a client connects/disconnects.
Not sure how that will play with a to-be-defined manual grouping.
But manual grouping isn't all that useful; most times one will need
to restart the processes, and doing so would require manual grouping
each time.  So better to use some automation, with labels.