mds-ss-beginning-of-statement: add heuristics
The function now searches for the start of a multi-part
statement (if/elif/else, try/catch).  It also returns
the statement number, which generally saves a call,
since that is what is needed.

Created mds-ss--regexp-for-statement to return a regexp
that matches the indentation of a statement, including
optional decoration.  This is used to move to a statement
at the same level of indentation as another (actually,
as part of the same extended statement: e.g. moving
from elif to if.

Added groups to constant mds-ss-statement-re to support
these changes.  Updated usage.