Fix breakpoints so that they can be set inside other known procedures.
This isn't ideal; the line-info for the location must be known,
which means that the procedure must have been entered.  Maybe this
limitation can be removed with a kernel modification.

LineInfo:-LookupStatement now returns an expression sequence.

The mds-ss-request function was modified to strip the trailing character,
which is assumed to be a newline.

mds-li-breakpoint now takes a CLEAR option; that avoids duplicating the procedure.