Partial fix for evan-and-prettyprint-prev.  Output isn't quite right.
The fix is really an improvement to the original and should be
backported to ss.  It goes directly to the preceding state, rather
than moving up a line and checking.  To implement this, the current
state is now passed to mds-li-update buffer and is stored as a
buffer-local in mds-li-state.  Possibly this shoud go into the client
so we can eliminate duplicate buffer-locals (from ss and li bufs).

The output isn't quite right; haven't tracked that down.  Looks
like an asynchronous issue.