Fix asynchronoous timing issue with mds-ss-request.  The fix is elsewhere.
The symptom was that mds-li-eval-and-prettyprint-prev would send a
request for the current rather than previous line.  My working
hypothesis is that the change of buffer caused by the rapid return of
the prompt (following a call to mds-ss-request) caused the point to
move.  To prevent this, the servicing of mds-ss-request now does not
cause an exit from _debugger, instead it just loops back to the wait
for another input.  To prevent another prompt from being printed,
debugger_readline now has a 'prompt' boolean parameter.