Handle local self-assignments by detecting their effect and correcting the statement number.
The effect is detected by noting the difference between the number of
statements indicated by the lineinfo data and the number of statements
in the interpreted procedure (as reported by debugopts(statcount).

Updated the documentation to indicate that this self-assignments
can now be used, provided they occur before regular assignments.