Display and hyperlink args on stack when mds-where is called.
Add a 'recusive debugging' tip to the documentation.
Describe the hyperlinked output of mds-where in the Output buffer section.

I don't recall why the arguments were not displayed; possibly
I was displaying the wrong ones at one time (the index into
the list is a bit weird) and so removed the feature.

A button with a right-pointing triangle is inserted to the left
of the argument list.  Clicking it causes the arguments to be
passed to the associated procedure.  This provides a convenient
method to recursively and repeatedly debug a procedure.

Moved and renamed mds-activate-addr-procname to the file
mds-out.el.

Insert 'require' statements into 'eval-when-compile'.
It makes sense, but invariably causes grief.